requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
aiosqlite
asyncpg
croniter
watchdog
requests
httpx
aiohttp
aio-pika
pika
redis
sseclient-py
networkx
cachetools
pandas
plotly
numpy
minio
sentence-transformers
dashscope
openai
vanna
chromadb
volcengine
mem0ai @ git+https://github.com/mem0ai/mem0
thespian
PyYAML
python-jose
cryptography
beautifulsoup4
psycopg2-binary
pymysql
pyodbc
cx_Oracle
neo4j
optuna
pyeda
scikit-learn
